some-fix

+++ app/controllers/items_controller.rb
@@ -34,9 +34,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if user_signed_in? && current_user.id == item.user_id
-      item.destroy
-    end
+    item.destroy if user_signed_in? && current_user.id == item.user_id
   end
 
   private
